Add comments about improving the execute method
In the current implementation, the PRINT instructions write their output to stdout. This PR adds comments how the Computer.execute method could be improved by returning the output as string, or a list of strings. In the latter case, each element could correspond to the output of a single PRINT instruction.